separated tests for mergability
