+++ src/tests/binary_ops.rs
@@ -1,0 +1,42 @@
+use engine::Engine;
+
+#[test]
+fn test_binary_ops() {
+    let mut engine = Engine::new();
+
+    if let Ok(result) = engine.eval::<i64>("10 % 4") {
+        assert_eq!(result, 2);
+    } else {
+        assert!(false);
+    }
+
+    if let Ok(result) = engine.eval::<i64>("10 << 4") {
+        assert_eq!(result, 160);
+    } else {
+        assert!(false);
+    }
+
+    if let Ok(result) = engine.eval::<i64>("10 >> 4") {
+        assert_eq!(result, 0);
+    } else {
+        assert!(false);
+    }
+
+    if let Ok(result) = engine.eval::<i64>("10 & 4") {
+        assert_eq!(result, 0);
+    } else {
+        assert!(false);
+    }
+
+    if let Ok(result) = engine.eval::<i64>("10 | 4") {
+        assert_eq!(result, 14);
+    } else {
+        assert!(false);
+    }
+
+    if let Ok(result) = engine.eval::<i64>("10 ^ 4") {
+        assert_eq!(result, 14);
+    } else {
+        assert!(false);
+    }
+}
