+++ src/tests/mod.rs
@@ -1,0 +1,24 @@
+mod arrays;
+mod binary_ops;
+mod bit_shift;
+mod bool_op;
+mod chars;
+mod comments;
+mod compound_equality;
+mod decrement;
+mod float;
+mod get_set;
+mod if_block;
+mod increment;
+mod internal_fn;
+mod looping;
+mod method_call;
+mod mismatched_op;
+mod not;
+mod number_literals;
+mod ops;
+mod string;
+mod unary_after_binary;
+mod unary_minus;
+mod var_scope;
+mod while_loop;
